fix unnessecary type conversion

+++ src/reduce.rs
@@ -40,14 +40,10 @@ pub const fn barrett_reduce(a: i16) -> i16 {
 #[test]
 fn test_barret_reduce() {
     for n in i16::MIN..i16::MAX {
-        let n_modq = n.wrapping_rem(Q as i16);
+        let n_modq = n.wrapping_rem(Q);
         let n_breduced = barrett_reduce(n);
         assert!(n_breduced < 3329 && n_breduced > -3329);
-        assert!(
-            n_modq == n_breduced
-                || n_modq - Q as i16 == n_breduced
-                || n_modq + Q as i16 == n_breduced
-        );
+        assert!(n_modq == n_breduced || n_modq - Q == n_breduced || n_modq + Q == n_breduced);
     }
 }
